move reference test function, change docs
