+++ torchmetrics/image/sam.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List
 
 from torch import Tensor
 from typing_extensions import Literal
@@ -55,7 +55,7 @@ class SpectralAngleMapper(Metric):
     preds: List[Tensor]
     target: List[Tensor]
     higher_is_better: bool = False
-    is_differentiable: bool = False
+    is_differentiable: bool = True
 
     def __init__(
         self,
